+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+
+### Fixed
+- import performance improved
+  ([issue #26](https://github.com/oboehm/ClazzFish/issues/26))
+
+
 ## [2.5.1] (12-Mar-2025)
 
 ### Fixed